test(saevm): refine shared pending-tx wait helper per review
