+++ vms/saevm/txgossip/BUILD.bazel
@@ -50,11 +50,13 @@ go_test(
         "//vms/saevm/saexec",
         "//vms/saevm/txgossip/txgossiptest",
         "@com_github_ava_labs_libevm//common",
+        "@com_github_ava_labs_libevm//core",
         "@com_github_ava_labs_libevm//core/rawdb",
         "@com_github_ava_labs_libevm//core/txpool",
         "@com_github_ava_labs_libevm//core/txpool/legacypool",
         "@com_github_ava_labs_libevm//core/types",
         "@com_github_ava_labs_libevm//eth",
+        "@com_github_ava_labs_libevm//event",
         "@com_github_ava_labs_libevm//libevm/ethapi",
         "@com_github_ava_labs_libevm//params",
         "@com_github_google_go_cmp//cmp",
