+++ vms/saevm/txgossip/txgossip_test.go
@@ -16,11 +16,13 @@ import (
 	"time"
 
 	"github.com/ava-labs/libevm/common"
+	"github.com/ava-labs/libevm/core"
 	"github.com/ava-labs/libevm/core/rawdb"
 	"github.com/ava-labs/libevm/core/txpool"
 	"github.com/ava-labs/libevm/core/txpool/legacypool"
 	"github.com/ava-labs/libevm/core/types"
 	"github.com/ava-labs/libevm/eth"
+	"github.com/ava-labs/libevm/event"
 	"github.com/ava-labs/libevm/libevm/ethapi"
 	"github.com/ava-labs/libevm/params"
 	"github.com/google/go-cmp/cmp"
@@ -143,7 +145,7 @@ func TestExecutorIntegration(t *testing.T) {
 	for _, tx := range signedTxs {
 		require.NoErrorf(t, s.Add(Transaction{tx}), "%T.Add()", s.set)
 	}
-	txgossiptest.WaitUntilPending(t, ctx, txgossiptest.PoolMempool(s.Pool), signedTxs...)
+	txgossiptest.WaitUntilPending(t, ctx, poolMempool{s.Pool}, signedTxs...)
 
 	t.Run("Iterate_after_Add", func(t *testing.T) {
 		require.Lenf(t, slices.Collect(s.Iterate), numTxs, "slices.Collect(%T.Iterate)", s.Set)
@@ -311,7 +313,7 @@ func TestP2PIntegration(t *testing.T) {
 
 			require.NoErrorf(t, send.Add(txViaGossip), "%T.Add()", send.Set)
 			require.NoErrorf(t, send.SendTx(ctx, txViaRPC.Transaction), "%T.SendTx()", send.Set)
-			txgossiptest.WaitUntilPending(t, ctx, txgossiptest.PoolMempool(send.Pool), txViaRPC.Transaction, txViaGossip.Transaction)
+			txgossiptest.WaitUntilPending(t, ctx, poolMempool{send.Pool}, txViaRPC.Transaction, txViaGossip.Transaction)
 
 			t.Run("confirm_setup", func(t *testing.T) {
 				for _, tx := range bothTxs {
@@ -432,4 +434,23 @@ func FuzzEffectiveGasTip(f *testing.F) {
 			t.Errorf("%T.effectiveGasTip(...) got %v; want %v", ltx, got, want)
 		}
 	})
+}
+
+// poolMempool adapts a raw [*txpool.TxPool] to [txgossiptest.Mempool], the
+// only shape these tests have access to without a geth RPC backend.
+type poolMempool struct {
+	pool *txpool.TxPool
+}
+
+func (m poolMempool) SubscribeNewTxsEvent(ch chan<- core.NewTxsEvent) event.Subscription {
+	return m.pool.SubscribeTransactions(ch, true /*reorgs ignored by legacypool*/)
+}
+
+func (m poolMempool) GetPoolTransactions() (types.Transactions, error) {
+	pendingByAddr, _ := m.pool.Content()
+	txs := make(types.Transactions, 0, len(pendingByAddr))
+	for _, list := range pendingByAddr {
+		txs = append(txs, list...)
+	}
+	return txs, nil
 }
