+++ vms/saevm/txgossip/txgossiptest/BUILD.bazel
@@ -10,7 +10,6 @@ go_library(
         "//utils/set",
         "@com_github_ava_labs_libevm//common",
         "@com_github_ava_labs_libevm//core",
-        "@com_github_ava_labs_libevm//core/txpool",
         "@com_github_ava_labs_libevm//core/types",
         "@com_github_ava_labs_libevm//event",
         "@com_github_stretchr_testify//require",
